+++ Sources/Core/Public/DIContainer.swift
@@ -6,6 +6,8 @@
 //  Copyright © 2016 Alexander Ivlev. All rights reserved.
 //
 
+import Foundation
+
 prefix operator *
 /// Short syntax for resolve.
 /// Using:
@@ -32,9 +34,43 @@ public final class DIContainer {
   internal let bundleContainer = BundleContainer()
   internal private(set) var resolver: Resolver!
   
-  // non thread safe!
-  internal var includedParts: Set<String> = []
-  internal var currentBundle: Bundle? = nil
+  ///MARK: Hierarchy
+  final class IncludedParts {
+    private var parts: Set<ObjectIdentifier> = []
+    private let mutex = PThreadMutex(recursive: ())
+    
+    func checkAndInsert(_ part: ObjectIdentifier) -> Bool {
+      return mutex.sync { parts.insert(part).inserted }
+    }
+  }
+  
+  final class BundleStack {
+    private let key = "ThreadSafeCurrentBundle\(UUID().uuidString)"
+    
+    var bundle: Bundle? { return stack?.last }
+    
+    func push(_ bundle: Bundle) {
+      if let stack = self.stack {
+        self.stack = stack + [bundle]
+      } else {
+        self.stack = [bundle]
+      }
+    }
+    
+    func pop() {
+      if let stack = self.stack {
+        self.stack = Array(stack.dropLast())
+      }
+    }
+    
+    private var stack: [Bundle]? {
+      set { return Thread.current.threadDictionary[key] = newValue }
+      get { return Thread.current.threadDictionary[key] as? [Bundle] }
+    }
+  }
+  
+  internal let includedParts = IncludedParts()
+  internal let bundleStack = BundleStack()
 }
 
 // MARK: - register
@@ -146,11 +182,13 @@ public extension DIContainer {
 public extension DIContainer {
   
   /// Validate the graph by checking various conditions. For faster performance, set false.
+	///
+	/// - Parameter checkGraphCycles: check cycles in the graph of heavy operation. So it can be disabled
   /// - Returns: true if validation success.
   @discardableResult
-  public func valid() -> Bool {
+	public func validate(checkGraphCycles isCheckGraphCycles: Bool = true) -> Bool {
     let components = componentContainer.components
-    return checkGraph(components) && checkGraphCycles(components)
+    return checkGraph(components) && (!isCheckGraphCycles || checkGraphCycles(components))
   }
 }
 
@@ -181,34 +219,34 @@ extension DIContainer {
         let candidates = resolver.findComponents(by: parameter.type, with: parameter.name, from: bundle)
         let filtered = resolver.removeWhoDoesNotHaveInitialMethod(components: candidates)
         
-        let correct = resolver.validate(components: filtered, for: parameter.type)
-        let success = correct || parameter.optional
+        let correct = 1 == filtered.count || parameter.type is IsMany.Type
+        let hasCachedLifetime = filtered.isEmpty && candidates.contains{ $0.lifeTime != .prototype }
+        let success = correct || parameter.optional || hasCachedLifetime
         successfull = successfull && success
-        
+
         // Log
         if !correct {
           if candidates.isEmpty {
-            plog(parameter, msg: "Not found component for \(description(type: parameter.type))")
+            plog(parameter, msg: "Not found component for \(description(type: parameter.type)) from \(component.info)")
           } else if filtered.isEmpty {
-            let allPrototypes = !candidates.contains{ $0.lifeTime != .prototype }
             let infos = candidates.map{ $0.info }
-            
-            if allPrototypes {
-              plog(parameter, msg: "Not found component for \(description(type: parameter.type)) that would have initialization methods. Were found: \(infos)")
+
+            if hasCachedLifetime {
+              log(.warning, msg: "Not found component for \(description(type: parameter.type)) from \(component.info) that would have initialization methods, but object can maked from cache. Were found: \(infos)")
             } else {
-              log(.warning, msg: "Not found component for \(description(type: parameter.type)) that would have initialization methods, but object can maked from cache. Were found: \(infos)")
+              plog(parameter, msg: "Not found component for \(description(type: parameter.type)) from \(component.info) that would have initialization methods. Were found: \(infos)")
             }
           } else if filtered.count >= 1 {
             let infos = filtered.map{ $0.info }
-            plog(parameter, msg: "Ambiguous \(description(type: parameter.type)) contains in: \(infos)")
+            plog(parameter, msg: "Ambiguous \(description(type: parameter.type)) from \(component.info) contains in: \(infos)")
           }
         }
       }
     }
-    
+		
     return successfull
   }
-  
+	
   fileprivate func checkGraphCycles(_ components: [Component]) -> Bool {
     var success: Bool = true
     
@@ -222,29 +260,30 @@ extension DIContainer {
             return true
           }
           
-          let components = stack.map{ $0.component.info }
+          let infos = stack.dropLast().map{ $0.component.info }
+          let short = infos.map{ "\($0.type)" }.joined(separator: " - ")
           
           let allInitials = !stack.contains{ !($0.initial && !$0.many) }
           if allInitials {
-            log(.error, msg: "You have a cycle: \(components) consisting entirely of initialization methods.")
+            log(.error, msg: "You have a cycle: \(short) consisting entirely of initialization methods. Full: \(infos)")
             return false
           }
           
           let hasGap = stack.contains{ $0.cycle || ($0.initial && $0.many) }
           if !hasGap {
-            log(.error, msg: "Cycle has no discontinuities. Please install at least one explosion in the cycle: \(components) using `injection(cycle: true) { ... }`")
+            log(.error, msg: "Cycle has no discontinuities. Please install at least one explosion in the cycle: \(short) using `injection(cycle: true) { ... }`. Full: \(infos)")
             return false
           }
           
           let allPrototypes = !stack.contains{ $0.component.lifeTime != .prototype }
           if allPrototypes {
-            log(.error, msg: "You cycle: \(components) consists only of object with lifetime - prototype. Please change at least one object lifetime to another.")
+            log(.error, msg: "You cycle: \(short) consists only of object with lifetime - prototype. Please change at least one object lifetime to another. Full: \(infos)")
             return false
           }
           
           let containsPrototype = stack.contains{ $0.component.lifeTime == .prototype }
           if containsPrototype {
-            log(.warning, msg: "You cycle: \(components) contains an object with lifetime - prototype. In some cases this can lead to an udesirable effect.")
+            log(.warning, msg: "You cycle: \(short) contains an object with lifetime - prototype. In some cases this can lead to an udesirable effect.  Full: \(infos)")
           }
           
           return true
@@ -262,10 +301,13 @@ extension DIContainer {
       
       func callDfs(by parameters: [MethodSignature.Parameter], initial: Bool, cycle: Bool) {
         for parameter in parameters {
-          let many = parameter.many
           let candidates = resolver.findComponents(by: parameter.type, with: parameter.name, from: bundle)
-          let filtered = resolver.removeWhoDoesNotHaveInitialMethod(components: candidates)
+          if candidates.isEmpty {
+            continue
+          }
           
+          let filtered = candidates.filter{ nil != $0.initial || ($0.lifeTime != .prototype && visited.contains($0)) }
+          let many = parameter.many
           for subcomponent in filtered {
             var stack = stack
             stack.append((subcomponent, initial, cycle, many))