+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# v3.0.4
+* bugfix: Improve validate graph cycle, and logs.
+
+# v3.0.3
+* bugfix: Support recursive inject into ViewControllers.
+* bugfix: Fix component bundle source.
+* bugfix: Fix valid method, for hard dependency graph.
+* Rename `valid` to `validate(checkGraphCycles:)`
+
+# v3.0.2
+* Changed hierarchy/default logic. Now component inside framework a upper priority than 'default'
+* Small improve code documentantion
+* Added generated documentation for the code
+
+# v3.0.1
+* Now thread safe supported for `append(framework:)`, `append(part:)`, `import()`
+* Now `append(framework:)` also call `import()`
+
 # v3.0.0
 * Added migration documentation.
 * Added code documentation.